Rename wasm package to `streams`
